+++ src/image_transform.cpp
@@ -27,8 +27,8 @@
 #include "miopen/common.hpp"
 #include "miopen/execution_context.hpp"
 #include "miopen/find_solution.hpp"
-#include "miopen/image_transform/invoke_params.hpp"
-#include "miopen/image_transform/problem_description.hpp"
+#include "miopen/image_transform/adjust_hue/invoke_params.hpp"
+#include "miopen/image_transform/adjust_hue/problem_description.hpp"
 #include "miopen/image_transform/solvers.hpp"
 #include "miopen/miopen.h"
 #include "miopen/image_transform.hpp"
