+++ src/include/miopen/image_transform/adjust_hue/invoke_params.hpp
