+++ src/include/miopen/image_transform/adjust_hue/problem_description.hpp
