+++ src/solver/image_transform/adjust_hue.cpp
@@ -27,8 +27,8 @@
 #include "miopen/conv_solution.hpp"
 #include "miopen/datatype.hpp"
 #include "miopen/errors.hpp"
-#include "miopen/image_transform/invoke_params.hpp"
-#include "miopen/image_transform/problem_description.hpp"
+#include "miopen/image_transform/adjust_hue/invoke_params.hpp"
+#include "miopen/image_transform/adjust_hue/problem_description.hpp"
 #include "miopen/image_transform/solvers.hpp"
 #include "miopen/invoke_params.hpp"
 #include "miopen/kernel_build_params.hpp"
